com-lib: Add pkg-config Autoconf program check
If pkg-config is not installed, attempting to use any of the PKG_CHECK
macros results in configure crashing in an ugly manner as these macros
are supplied in the pkg-config package. So, introduce a program check
for pkg-config before these macros are used so that configure can error
out gracefully.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>