srv: com: Add id request and use for each request
When using ssh the daemon always sees the other end as localhost due to
the connection from the end of the ssh tunnel to the daemon being local.
So add an id request type which must be the first request of any
discreet session. Now the daemon knows who is sending the request.

Whilst not considered a great risk, it should be noted that someone
could write a malicious program using the swoc libraries to send
messages with a spoofed host name or IP. However, the daemon does try to
validate the information by doing a reverse lookup.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>
(cherry picked from commit 8b3eca060b44436fdad97e571bc4d874f617ceed)
Signed-off-by: Mark Grant <m.grant.prg@gmail.com>