at: Register docs with scrollkeeper only if ATONLY
If documents are registered with scrollkeeper as part of an AT
installation then conflicts will arise if the documents have already
been installed with the package manager. This is because the DOCID will
be identical.

To avoid these conflicts only register documents with scrollkeeper if
the enable-atonly option has been passed to configure.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>