at: Add support for gcc analyzer option
Add a command line option to configure to optionally enable support for
using gcc's analyzer flag. Configure retrieves the flags necessary and
passes them to the makefiles. If the gcc version does not support the
analyzer, which was first available in version 10, then the string passed
to the makefiles is empty and the whole thing becomes a no-op.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>