at: doc: Remove portability constraints for Automake
The doxygen documentation depends on it's own source files and the C
source files. It is not practical to list all the dependencies manually.
To perform it automatically requires the use of GNU Make extensions so
portability warnings must be disabled. Then find can be used in the
Makefile.am

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>