srv-prg: IP / Hostname ID matching spamming the logs
The ID verification handshake between client or server and daemon
always logs failure and success. This causes multiple entries in
syslog as seen below:-

2024-09-17T09:51:51.127672+01:00 hermes swocserverd: Host and IP matched - leonidas.mgent.home 192.168.62.42
2024-09-17T09:51:51.142827+01:00 hermes swocserverd: Host and IP matched - leonidas.mgent.home 192.168.62.42
.....

In particular a message will be emitted to the syslog every poll
interval during a wait function, potentially leading to a long string of
such messages.

Logging ID success is not very interesting as it will always be
followed by the user-requested CLA being processed.

Solve this by only logging success when the daemon is run with the
--debug CLA.

GitHub: Fixes #13
Signed-off-by: Mark Grant <m.grant.prg@gmail.com>